refactor: Product 가격 필드에 Money 값 객체 적용

+++ product-server/src/main/java/com/drf/product/model/request/ProductUpdateRequest.java
@@ -15,7 +15,7 @@ public record ProductUpdateRequest(
         String name,
 
         @Min(0)
-        Integer price,
+        Long price,
 
         @Size(min = 1)
         String description,
